feat: 미참조 v1 로그인 엔드포인트 제거
