+++ bottlenote-product-api/src/main/java/app/bottlenote/user/controller/OauthController.java
@@ -6,10 +6,10 @@ import app.bottlenote.global.annotation.SecurityPolicy;
 import app.bottlenote.global.data.response.GlobalResponse;
 import app.bottlenote.user.config.OauthConfigProperties;
 import app.bottlenote.user.dto.request.BasicLoginRequest;
-import app.bottlenote.user.dto.request.OauthRequest;
 import app.bottlenote.user.dto.request.TokenVerifyRequest;
 import app.bottlenote.user.dto.response.OauthResponse;
 import app.bottlenote.user.dto.response.TokenItem;
+import app.bottlenote.user.service.AuthService;
 import app.bottlenote.user.service.OauthService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -31,34 +31,22 @@ import org.springframework.web.bind.annotation.RestController;
 @SecurityPolicy(auth = PUBLIC)
 public class OauthController {
   private static final String REFRESH_TOKEN_HEADER_PREFIX = "refresh-token";
+  private final AuthService authService;
   private final OauthService oauthService;
   private final OauthConfigProperties configProperties;
-
-  @PostMapping("/login")
-  public ResponseEntity<?> executeOauthLogin(
-      @RequestBody @Valid OauthRequest oauthReq, HttpServletResponse response) {
-    TokenItem token = oauthService.login(oauthReq);
-    setRefreshTokenInCookie(response, token.refreshToken());
-    return GlobalResponse.ok(OauthResponse.of(token.accessToken()));
-  }
 
   @PostMapping("/reissue")
   public ResponseEntity<?> reissueOauthToken(
       HttpServletRequest request, HttpServletResponse response) {
-
     String refreshToken = request.getHeader(REFRESH_TOKEN_HEADER_PREFIX);
-
-    TokenItem token = oauthService.refresh(refreshToken);
-
+    TokenItem token = authService.reissue(refreshToken);
     setRefreshTokenInCookie(response, token.refreshToken());
-
     return GlobalResponse.ok(OauthResponse.of(token.accessToken()));
   }
 
   @PutMapping("/token/verify")
   public ResponseEntity<?> verifyToken(@RequestBody @Valid TokenVerifyRequest token) {
-    final String message = oauthService.verifyToken(token.token());
-    return GlobalResponse.ok(message);
+    return GlobalResponse.ok(authService.verifyToken(token.token()));
   }
 
   @PostMapping("/restore")
@@ -68,15 +56,12 @@ public class OauthController {
   }
 
   private void setRefreshTokenInCookie(HttpServletResponse response, String refreshToken) {
-    final int COOKIE_EXPIRE_TIME = 14 * 24 * 60 * 60;
     final int cookieExpireTime = configProperties.getCookieExpireTime();
-    log.debug(
-        "cookie basic expire time : {} properties time :{}", COOKIE_EXPIRE_TIME, cookieExpireTime);
     Cookie cookie = new Cookie(REFRESH_TOKEN_HEADER_PREFIX, refreshToken);
     cookie.setHttpOnly(true);
     cookie.setSecure(true);
     cookie.setPath("/");
-    cookie.setMaxAge(COOKIE_EXPIRE_TIME);
+    cookie.setMaxAge(cookieExpireTime);
     response.addCookie(cookie);
   }
 }
